apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.12.5
pydantic-settings==2.9.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.18
stripe==11.4.0
redis==5.2.1
structlog==24.4.0
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
ruff==0.8.6
mypy==1.13.0
email-validator==2.2.0
croniter==3.0.3
pyotp==2.9.0
openpyxl==3.1.5
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
sentry-sdk[fastapi]==2.20.0
